Rename notice component to PascalCase
