+++ app/javascript/article-form/components/Notice.jsx
